posix: c_lib_ext: fnmatch: fix escape-oriented regression
A regression in 936d0278bdfa773390469aa102a3de004f355514 introduced a
subtle bug in the way that escaped expressions were handled.

The regression originated with the assumption that test data (originally
adapted from a 3rd-party testsuite) was correct when it was in fact
flawed.

Specifically, `fnmatch("[[?*\\]", "\\", 0)` should fail, since the "\\"
sequence (a single backslash after compilation) escapes the following ']'
character, thus leaving the bracket expression incomplete. Since the
bracket expression (and therefore the whole pattern) is erroneous, the
call should return `FNM_NOMATCH` to indicate failure rather than 0 to
indicate success.

Added new test cases from #98827 and some commentary for subsequent
reviewers.

This change does not complete #55186 but is related to it.

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>
Signed-off-by: Jukka Rissanen <jukka.rissanen@nordicsemi.no>
Signed-off-by: Keith Packard <keithp@keithp.com>