tests: posix: c_lib_ext: add tests for strfmon()
Add a test for strfmon().

The values in the testsuite for "C" and "en_US.UTF-8" have been validated
to work with Zephyr and glibc.

It's unclear whether "en_US.UTF-8" or any other internationalization will
every be used within Zephyr, so the test-suite, as well as the
implementation within Zephyr, have been designed for only the "C" locale.

In the future, if additional locales are required, then they can be added
easily via Kconfig.

In the implementation, the `stfmon_one()` function should be all that
needs to be reimplemented on a per-locale basis.

Signed-off-by: Chris Friedt <chris@fr4.co>