tests: posix: c_lib_ext: getopt: rewrite testsuite based on spec
Previously, the getopt testsuite included tests for non-standard features
and would not pass even with a conformant implementation.

Rewrite the testsuite to to check for a conformance.

Remove checks for functionality that do not exist in conformant
implementations.

Additionally, given that `getopt_long()`, and `getopt_long_only()` are not
specified by POSIX, remove those from the testsuite.

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>