posix: c_lib_ext: getopt: reimplement getopt() to be mostly conformant
Previously, Zephyr's implementation of `getopt()` was not conformant to
the POSIX specification due to non-standard `getopt_init()`, state-saving
functions, and so on.

Reimplement `getopt()` so that it is conformant to the POSIX
specification.

The minor exception to conformance (i.e. a deviation) is that we use
`opterr` to save state for multi-option arguments like "-abc".

Additionally, remove `getopt_long()` and `getopt_long_only()` as those are
GNU extensions and are not required for the POSIX_C_LIB_EXT Option Group.

Signed-off-by: Chris Friedt <chris@fr4.co>