posix: c_lib_ext: strfmon: check if right-precision is CHAR_MAX
When `frac_digits` (or really any of the `struct lconv` entries) is
equal to `CHAR_MAX`, it indicates that the value is not available in the
current locale.

We currently only rely on that field for `right_precision`, so if the
value is not available in the current locale, "C", then set right
precision to a rational default.

When a rational default is not used, it's possible for a divide-by-zero
to occur, because shifting an integer scaling factor by `CHAR_MAX` will
overflow a 64-bit value.

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>