tests: posix: c_lib_ext: compile even when chosen is not present
Previously, if the DT chosen node `zephyr,entropy` was not configured, the
getentropy.c test would not compile.

Maybe one day we will have `DT_CHOSEN_OR()` and `DT_CHOSEN_OR_NULL()`,
which would presumably percolate `NULL` through to
`DEVICE_DT_GET_OR_NULL()` so that it could be a one-liner, but for now,
make a conditional declaration on `DT_HAS_CHOSEN()`.

Signed-off-by: Chris Friedt <chris@fr4.co>