posix: pthread: elastipool for object pools
Use the newly introduced elastipool elastic object pool for management
of

* barrier
* cond
* key
* mutex
* pthread
* rwlock
* spinlock

objects.

This change only changes the storage API used for those areas from
an array + bitarray to an elastipool, which is effectively the same thing,
but without spinlock synchronization. The main reason for doing so is that
spinlocks cannot be used from userspace code, so synchronization likely
does not work as expected in that case. Instead of the built-in spinlock
used with the bitarray API, use an external sys_sem for locking, which
may be used from userspace code.

This does not enable heap allocation of objects, as the min / max values
remain equal to their original Kconfig limits. Since min == max, all
allocations are static / guaranteed as part of the statically allocated
"slab" storage.

One "drive-by" update was made - specifically, to use atomics in the
POSIX spinlock implementation, instead of kernel spinlocks.

Signed-off-by: Chris Friedt <chris@fr4.co>