posix: eventfd: deadlock fixes
When eventfd is used through read(2) and write(2), the mutex is
already locked from the fdtable implementation. So we remove the
usage of the mutex from the zvfs_eventfd_*_op functions, as it is
already managed by fdtable.

However, when zvfs_eventfd_{read,write} are used, no fdtable layer
is used and we shuld call the _op function with the mutex locked
(the same behavior as with fdtable), so these functions should
manage the mutex. We add it there.

Signed-off-by: Marco Casaroli <marco.casaroli@gmail.com>
Signed-off-by: Chris Friedt <chris@fr4.co>