posix: eventfd: ensure interoperability of the eventfd tests and impl
It's desireable to ensure that we test `eventfd()` behaviour against the
Linux implementation for interoperability. In order to do so, local to
the testsuite, several `TC_PROVIDES` macros were added to `NATIVE_LIBC`
for everything ranging from DEVICE_IO to PTHREADS, so that the testsuite
could be run without source modification.

Interestingly, we learned that Linux eventfd does not support F_GETFL or
F_SETFL ioctls, and also, that a blocking read operation will not return
with an error when the underlying file descriptor is closed, mid-read.
As a result, the `test_read_then_close_block` test case was removed.

No POSIX Option Groups were harmed in the making of this testsuite.

However, C++ hash maps needed to be disabled due to some random Kconfigs
dependency cycle relating to FULL_LIBC.

Signed-off-by: Chris Friedt <chris@fr4.co>