posix: threads: move impl to sys/thread.h and refactor thrd impl
Refactor the implementation of POSIX threads out of lib/posix, and into
lib/os as the sys/thread.h API which mainly serves to complement the
k_thread (k_mutex, k_condvar) APIs.

This change was required in order to remove POSIX from the ISO C threads
equation and simultaneously enables us to shrink the POSIX API to the
bare minimum required.

Signed-off-by: Chris Friedt <chris@fr4.co>